7791: basic refactoring to export some avm test helpers
