+++ yarn-project/bb-prover/src/test/index.ts
@@ -1,2 +1,3 @@
 export * from './test_circuit_prover.js';
 export * from './test_verifier.js';
+export * from './test_avm.js';
